combined krpc and integrator sim into one file, passing unit tests
