+++ gcherry/tests/test_rocket_ode.py
@@ -3,7 +3,7 @@ import math
 import numpy as np
 
 from gcherry.rk4 import rk4
-from gcherry.integrator_sim import rocket_ode
+from gcherry.sim_interface import rocket_ode
 
 
 class TestRocketOde(unittest.TestCase):
